+++ server/src/config.ts
@@ -36,7 +36,6 @@ export interface VLSConfig {
         user: string;
         vetur: string;
       };
-      stylusSeparator: boolean;
     };
     grammar: {
       customBlocks: { [lang: string]: string };
@@ -91,6 +90,7 @@ export interface VLSFullConfig extends VLSConfig {
   typescript?: any;
   prettier?: any;
   stylusSupremacy?: any;
+  languageStylus?: any;
 }
 
 export function getDefaultVLSConfig(): VLSFullConfig {
@@ -112,8 +112,7 @@ export function getDefaultVLSConfig(): VLSFullConfig {
           workspace: '💼',
           user: '🗒️',
           vetur: '✌'
-        },
-        stylusSeparator: false
+        }
       },
       grammar: {
         customBlocks: {}
@@ -158,7 +157,8 @@ export function getDefaultVLSConfig(): VLSFullConfig {
       format: {}
     },
     emmet: {},
-    stylusSupremacy: {}
+    stylusSupremacy: {},
+    languageStylus: {}
   };
 }
 
@@ -249,7 +249,9 @@ export async function getVeturFullConfig(
     })
     .sort((a, b) => {
       const r = getPathDepth(b.root, '/') - getPathDepth(a.root, '/');
-      if (r !== 0) { return r; }
+      if (r !== 0) {
+        return r;
+      }
       return b.root.length - a.root.length;
     });